Last of naming updates, view creation, and bug fixes
