+++ db/scripts/3_create_l2_tables.sql
@@ -4,20 +4,22 @@
 CREATE TABLE optimistic_canonical_chain_batch (
   id BIGSERIAL NOT NULL,
   batch_number BIGINT NOT NULL,
-  submitted_l1_tx_hash character(66) DEFAULT NULL,
-  tx_status TEXT NOT NULL DEFAULT 'QUEUED',
-  root_status TEXT NOT NULL DEFAULT 'QUEUED',
+  submitted_tx_batch_l1_tx_hash character(66) DEFAULT NULL,
+  submitted_root_batch_l1_tx_hash character(66) DEFAULT NULL,
+  tx_batch_status TEXT NOT NULL DEFAULT 'QUEUED',
+  root_batch_status TEXT NOT NULL DEFAULT 'QUEUED',
   created TIMESTAMP NOT NULL DEFAULT NOW(),
   PRIMARY KEY (id),
   UNIQUE (batch_number)
 );
-CREATE INDEX optimistic_canonical_chain_batch_tx_status_idx ON optimistic_canonical_chain_batch USING btree (tx_status);
-CREATE INDEX optimistic_canonical_chain_batch_root_status_idx ON optimistic_canonical_chain_batch USING btree (root_status);
+CREATE INDEX optimistic_canonical_chain_batch_tx_status_idx ON optimistic_canonical_chain_batch USING btree (tx_batch_status);
+CREATE INDEX optimistic_canonical_chain_batch_root_status_idx ON optimistic_canonical_chain_batch USING btree (root_batch_status);
 
 CREATE TABLE l2_tx_output (
   id BIGSERIAL NOT NULL,
   l1_rollup_tx_id BIGINT DEFAULT NULL,
   occ_batch_number BIGINT DEFAULT NULL,
+  occ_batch_index INT DEFAULT NULL,
   block_number BIGINT NOT NULL,
   block_timestamp BIGINT NOT NULL,
   tx_index INT NOT NULL,
