+++ packages/rollup-core/test/app/l2-node-service.spec.ts
@@ -49,7 +49,7 @@ const sender: string = Wallet.createRandom().address
 const target: string = Wallet.createRandom().address
 const calldata: string = keccak256FromUtf8('calldata')
 const rollupTx: RollupTransaction = {
-  batchIndex: 1,
+  indexWithinSubmission: 1,
   gasLimit,
   nonce,
   sender,
@@ -69,7 +69,7 @@ const sender2: string = Wallet.createRandom().address
 const target2: string = Wallet.createRandom().address
 const calldata2: string = keccak256FromUtf8('calldata 2')
 const rollupTx2: RollupTransaction = {
-  batchIndex: 2,
+  indexWithinSubmission: 2,
   gasLimit: gasLimit2,
   nonce: nonce2,
   sender: sender2,
@@ -91,6 +91,8 @@ const deserializeBlockBatches = (serialized: string): BlockBatches => {
       case 'gasLimit':
       case 'nonce':
       case 'batchIndex':
+      case 'indexWithinSubmission':
+      case 'queueIndex':
       case 'l1BlockNumber':
       case 'l1Timestamp':
       case 'queueOrigin':
