+++ packages/rollup-core/test/app/log-handlers.spec.ts
@@ -96,11 +96,11 @@ class MockDataService extends DefaultDataService {
     super(undefined)
   }
 
-  public async createNextL1ToL2Batch(): Promise<number> {
+  public async queueNextL1ToL2GethSubmission(): Promise<number> {
     return ++this.createdL1ToL2Batches
   }
 
-  public async createNextSafetyQueueBatch(): Promise<number> {
+  public async createNextSafetyQueueGethSubmission(): Promise<number> {
     return ++this.createdSafetyQueueBatches
   }
 
@@ -189,7 +189,7 @@ describe('Log Handlers', () => {
       QueueOrigin.L1_TO_L2_QUEUE,
       'Queue Origin mismatch'
     )
-    received.batchIndex.should.equal(0, 'Batch index mismatch')
+    received.indexWithinSubmission.should.equal(0, 'Batch index mismatch')
     received.sender.should.equal(l.address, 'Sender mismatch')
     remove0x(received.l1MessageSender).should.equal(
       sender,
@@ -242,7 +242,7 @@ describe('Log Handlers', () => {
       QueueOrigin.SAFETY_QUEUE,
       'Queue Origin mismatch'
     )
-    received.batchIndex.should.equal(0, 'Batch index mismatch')
+    received.indexWithinSubmission.should.equal(0, 'Batch index mismatch')
     remove0x(received.sender).should.equal(
       remove0x(wallet.address),
       'Sender mismatch'
@@ -329,7 +329,7 @@ describe('Log Handlers', () => {
         QueueOrigin.SEQUENCER,
         'Queue Origin mismatch'
       )
-      received.batchIndex.should.equal(i, 'Batch index mismatch')
+      received.indexWithinSubmission.should.equal(i, 'Batch index mismatch')
       remove0x(received.sender).should.equal(
         remove0x(wallet.address),
         'Sender mismatch'
